Protect prior notification PDF download & check
