+++ backend/src/test/kotlin/fr/gouv/cnsp/monitorfish/infrastructure/api/bff/PriorNotificationControllerUTests.kt
@@ -6,9 +6,7 @@ import com.nhaarman.mockitokotlin2.anyOrNull
 import fr.gouv.cnsp.monitorfish.config.SentryConfig
 import fr.gouv.cnsp.monitorfish.domain.entities.facade.SeafrontGroup
 import fr.gouv.cnsp.monitorfish.domain.entities.logbook.LogbookMessagePurpose
-import fr.gouv.cnsp.monitorfish.domain.entities.prior_notification.ManualPriorNotificationComputedValues
-import fr.gouv.cnsp.monitorfish.domain.entities.prior_notification.PriorNotificationState
-import fr.gouv.cnsp.monitorfish.domain.entities.prior_notification.PriorNotificationStats
+import fr.gouv.cnsp.monitorfish.domain.entities.prior_notification.*
 import fr.gouv.cnsp.monitorfish.domain.use_cases.prior_notification.*
 import fr.gouv.cnsp.monitorfish.domain.utils.PaginatedList
 import fr.gouv.cnsp.monitorfish.fakers.PriorNotificationFaker
@@ -26,8 +24,7 @@ import org.springframework.context.annotation.Import
 import org.springframework.http.MediaType
 import org.springframework.test.web.servlet.MockMvc
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*
-import org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath
-import org.springframework.test.web.servlet.result.MockMvcResultMatchers.status
+import org.springframework.test.web.servlet.result.MockMvcResultMatchers.*
 import java.time.ZonedDateTime
 
 @Import(SentryConfig::class)
@@ -48,6 +45,9 @@ class PriorNotificationControllerUTests {
 
     @MockBean
     private lateinit var getPriorNotification: GetPriorNotification
+
+    @MockBean
+    private lateinit var getPriorNotificationPdfDocument: GetPriorNotificationPdfDocument
 
     @MockBean
     private lateinit var getPriorNotifications: GetPriorNotifications
@@ -392,6 +392,75 @@ class PriorNotificationControllerUTests {
             // Then
             .andExpect(status().isOk)
             .andExpect(jsonPath("$.reportId", equalTo(fakePriorNotification.reportId)))
+    }
+
+    @Test
+    fun `getPdf Should get the PDF of a prior notification`() {
+        val dummyPdfContent = """
+            %PDF-1.4
+            1 0 obj
+            << /Type /Catalog /Pages 2 0 R >>
+            endobj
+            2 0 obj
+            << /Type /Pages /Kids [3 0 R] /Count 1 >>
+            endobj
+            3 0 obj
+            << /Type /Page /Parent 2 0 R /MediaBox [0 0 612 792] /Contents 4 0 R >>
+            endobj
+            4 0 obj
+            << /Length 55 >>
+            stream
+            BT
+            /F1 24 Tf
+            100 700 Td
+            (Hello, World!) Tj
+            ET
+            endstream
+            endobj
+            xref
+            0 5
+            0000000000 65535 f
+            0000000010 00000 n
+            0000000062 00000 n
+            0000000113 00000 n
+            0000000218 00000 n
+            trailer
+            << /Size 5 /Root 1 0 R >>
+            startxref
+            291
+            %%EOF
+        """.trimIndent().toByteArray()
+
+        // Given
+        given(getPriorNotificationPdfDocument.execute("REPORT_ID", false))
+            .willReturn(
+                PdfDocument(
+                    reportId = "REPORT_ID",
+                    source = PriorNotificationSource.LOGBOOK,
+                    generationDatetimeUtc = ZonedDateTime.now(),
+                    pdfDocument = dummyPdfContent,
+                ),
+            )
+
+        // When
+        api.perform(get("/bff/v1/prior_notifications/REPORT_ID/pdf"))
+            // Then
+            .andExpect(status().isOk)
+            .andExpect(content().contentType(MediaType.APPLICATION_PDF))
+            .andExpect(content().bytes(dummyPdfContent))
+    }
+
+    @Test
+    fun `getPdfExistence Should get the PDF existence of a prior notification`() {
+        // Given
+        given(getPriorNotificationPdfDocument.execute("REPORT_ID", true)).willReturn(null)
+
+        // When
+        api.perform(get("/bff/v1/prior_notifications/REPORT_ID/pdf/exist"))
+            // Then
+            .andExpect(status().isOk)
+            .andExpect(content().contentType(MediaType.APPLICATION_JSON))
+            .andExpect(jsonPath("$.status", equalTo("NO_CONTENT")))
     }
 
     @Test
